Add files via upload
Airbnb is now one of the most successful modes of sharing economy and millions of homeowners and businesses benefits from listing on Airbnb. Therefore, analysing and exploring the factors that lead to a successful listing is of great value for the investors and hosts. This is the main reason drives the research. Besides, there are many metrics in the Airbnb listing, which is also a good sample to analyse the relationship between various factors and price modelling.